backend/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.31.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.6.0
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
